+++ src/main/java/uk/gov/hmcts/reform/pdf/service/appinsights/AppInsightsEvent.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pdf.service;
+package uk.gov.hmcts.reform.pdf.service.appinsights;
 
 public enum AppInsightsEvent {
