+++ src/test/java/uk/gov/hmcts/reform/pdf/service/endpoint/GetWelcomeTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.pdf.generator.HTMLToPDFConverter;
-import uk.gov.hmcts.reform.pdf.service.AppInsightsEventTracker;
+import uk.gov.hmcts.reform.pdf.service.appinsights.AppInsightsEventTracker;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
